Init composer-ised RDF-knowing burrow
