+++ templates/index.php
@@ -1,0 +1,107 @@
+<!doctype html>
+<html>
+  <head>
+    <title>Burrow</title>
+    <link rel="stylesheet" type="text/css" href="https://apps.rhiaro.co.uk/css/normalize.min.css" />
+    <link rel="stylesheet" type="text/css" href="https://apps.rhiaro.co.uk/css/main.css" />
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+  </head>
+  <body>
+    <main class="w1of2 center">
+      <h1>Burrow</h1>
+      
+      <?if(isset($errors)):?>
+        <div class="fail">
+          <?foreach($errors as $key=>$error):?>
+            <p><strong><?=$key?>: </strong><?=$error?></p>
+          <?endforeach?>
+        </div>
+      <?endif?>
+      
+      <?if(isset($result)):?>
+        <div>
+          <p>The response from you your micropub endpoint:</p>
+          <code><?=$endpoint?></code>
+          <?if($result['code'] != "201"):?>
+            <p class="fail">Nothing created, error code <strong><?=$result['code']?></strong></p>
+          <?else:?>
+            <p class="win">Post created.. <strong><?=$result['location'][0]?></strong></p>
+          <?endif?>
+          <pre>
+            <? var_dump($result['response']); ?>
+          </pre>
+        </div>
+      <?endif?>
+      
+      <form method="post" role="form" id="checkin" class="align-center">
+        <?foreach($locations as $location):?>
+          <p><button class="neat inner color3-bg" style="border: none; width: 100%; padding: 1em;<?=isset($location["color"]) ? " background-color: ".$location["color"].";" : ""?>" type="submit" value="<?=$location['id']?>" name="location"><?=$location['name']?></button></p>
+        <?endforeach?>
+        <p>
+          <select name="year" id="year">
+            <option value="2019" selected>2019</option>
+            <option value="2018">2018</option>
+          </select>
+          <select name="month" id="month">
+            <?for($i=1;$i<=12;$i++):?>
+              <option value="<?=date("m", strtotime("2016-$i-01"))?>"<?=(date("n") == $i) ? " selected" : ""?>><?=date("M", strtotime("2016-$i-01"))?></option>
+            <?endfor?>
+          </select>
+          <select name="day" id="day">
+            <?for($i=1;$i<=31;$i++):?>
+              <option value="<?=date("d", strtotime("2016-01-$i"))?>"<?=(date("j") == $i) ? " selected" : ""?>><?=date("d", strtotime("2016-01-$i"))?></option>
+            <?endfor?>
+          </select>
+          <input type="text" name="time" id="time" value="<?=date("H:i:s")?>" />
+          <input type="text" name="zone" id="zone" value="<?=date("P")?>" />
+        </p>
+      </form>
+      
+      <div class="color3-bg inner">
+        <?if(isset($_SESSION['me'])):?>
+          <p class="wee">You are logged in as <strong><?=$_SESSION['me']?></strong> <a href="?logout=1">Logout</a></p>
+        <?else:?>
+          <form action="https://indieauth.com/auth" method="get" class="inner clearfix">
+            <label for="indie_auth_url">Domain:</label>
+            <input id="indie_auth_url" type="text" name="me" placeholder="yourdomain.com" />
+            <input type="submit" value="signin" />
+            <input type="hidden" name="client_id" value="https://rhiaro.co.uk" />
+            <input type="hidden" name="redirect_uri" value="<?=$base?>" />
+            <input type="hidden" name="state" value="<?=$base?>" />
+            <input type="hidden" name="scope" value="post" />
+          </form>
+        <?endif?>
+        
+        <h2>Customise</h2>
+        <h3>Locations</h3>
+        <?if(isset($locations_source)):?>
+          <p class="wee">Your locations are from <strong><?=$locations_source?></strong> <a href="?reset=1">Reset</a></p>
+        <?else:?>
+          <form method="post" class="inner wee clearfix">
+            <p>If you have or know a webpage with an archive of locations/venues you'd like on this list, enter the URL here. Currently can only read AS2 Collections, unpaged, returned in JSON or JSON-LD.</p>
+            <label for="locations_source">URL of a list of locations:</label>
+            <input id="locations_source" name="locations_source" value="https://rhiaro.co.uk/locations" />
+            <input type="submit" value="Fetch" />
+          </form>
+        <?endif?>
+        <h3>Post...</h3>
+        <form method="post" class="inner wee clearfix">
+          <select name="posttype">
+            <option value="as2" selected>AS2 JSON</option>
+            <option value="mp" disabled>Micropub (form-encoded)</option>
+            <option value="mp" disabled>Micropub (JSON)</option>
+            <option value="ttl" disabled>Turtle</option>
+          </select>
+          <input type="submit" value="Save" />
+        </form>
+      </div>
+    </main>
+    <footer class="w1of2 center">
+      <p><a href="https://github.com/rhiaro/burrow">Code</a> | <a href="https://github.com/rhiaro/burrow/issues">Issues</a>
+      <?if(isset($_SESSION['access_token'])):?>
+        | <a href="https://apps.rhiaro.co.uk/burrow?token=<?=$_SESSION['access_token']?>">Quicklink</a>
+      <?endif?>
+      </p>
+    </footer>
+  </body>
+</html>
